requirements.txt
streamlit
pandas
scikit-learn
mysql-connector-python
python-dotenv
speechrecognition